Update README.md
When I tried to build this project, I didn't found "Driver Files" in the menu of "Solution Explorer". It was necessary to change the platform toolset of the projects to "WindowsKernelModeDriver10.0" or "WindowsUserModeDriver10.0". The steps I added should be added in the introduction of building this project. Hope this could help others who meet the same problems when building the projects.